+++ src/Reader.php
@@ -304,7 +304,7 @@ class Reader extends AbstractCsv implements TabularDataReader, JsonSerializable
         }
 
         if ($header !== array_unique(array_filter($header, 'is_string'))) {
-            throw SyntaxError::dueToInvalidHeaderContent();
+            throw SyntaxError::dueToInvalidHeaderContent($header);
         }
 
         return $header;